+++ server/src/routes/poll.test.ts
@@ -6,12 +6,12 @@ vi.mock("~/repository/poll-repository", () => ({
   },
 }));
 
-vi.mock("~/services/poll-response", () => ({
+vi.mock("~/services/poll-results", () => ({
   getPollResults: vi.fn(),
 }));
 
 const { pollRepository } = await import("~/repository/poll-repository");
-const { getPollResults } = await import("~/services/poll-response");
+const { getPollResults } = await import("~/services/poll-results");
 const { pollRoutes: rawRoutes } = await import("./poll");
 
 import { withResolvePrincipal } from "~/__tests__/helpers/test-app";
